Function to create indicator results tables

This function performs an exact binomial test on the data for the indicator of interest and provides an estimate of the indicator proportion and the corresponding 95% confidence interval. This assumes that the probability weights of selection of an respondent are equal (given that sample has been taken using probability proportional to population size, this assumption holds true unless otherwise stated).

create_table(df, var)

Arguments

df
A recoded data.frame of indicators

var
Indicator to create a results table for

Value

A data.frame of indicator results with 95% confidence limits for each state in total and disaggregated by urban, rural and hard-to-reach areas.

Examples

## Recode anthro indicators
temp <- recode_anthro(df = create_canthro(df = childAnthro, x = anthroDF))

#> ================================================================================
#> ================================================================================
#> ================================================================================

create_table(df = temp, var = "global.haz")

#>                          Estimate           95% LCL           95% UCL
#> 1                         State 1           State 1           State 1
#> 2         Rural 0.166666666666667 0.142808386358973 0.192744366757418
#> 3         Urban 0.159235668789809 0.127356759096701 0.195464214675862
#> 4 Hard-to-reach 0.432478632478632 0.391906695118041 0.473735649495882
#> 5         Total 0.297752808988764 0.275436329054198 0.320812825832608
#>            Estimate           95% LCL           95% UCL
#> 1           State 2           State 2           State 2
#> 2  0.18657937806874 0.156447367667534 0.219776209750583
#> 3 0.141025641025641 0.114667289132968 0.170833410232544
#> 4 0.380952380952381 0.340809349529312  0.42234273821193
#> 5  0.23196448390677 0.212647728567796 0.252157694505986