Function to create indicator results tables

Function to create indicator results tables

create_unweighted_table(df, ppiDF = recode_ppi(df = hh), vars, digits = NULL)

Arguments

df
A recoded data.frame of indicators

ppiDF
A data.frame produced when recode_ppi is applied to the hh dataset of the Myanmar MCCT Programme Evaluation

vars
Indicator/s to create a results table for

digits
Integer value for number of digits to round off values. Default is NULL for no rounding off.

Value

A data.frame of indicator results for each state in total and disaggregated by urban, rural and hard-to-reach areas and by wealth quintile.

Examples

## Recode wash indicators
washDF  <- recode_wash(df = hh)

create_unweighted_table(df = washDF,
                        vars = c("summer1", "summer6", "summer5",
                                 "rain1", "rain6", "rain5",
                                 "winter1", "winter6", "winter5"))

#>    state        strata   summer1    summer6     summer5     rain1      rain6
#> 1  Kayah         Rural 0.7460815 0.07836991 0.175548589 0.8025078 0.10658307
#> 3  Kayah         Urban 0.8953488 0.04941860 0.055232558 0.8895349 0.05523256
#> 5  Kayah Hard-to-reach 0.6383562 0.16986301 0.191780822 0.6876712 0.16986301
#> 7  Kayah    Wealthiest 0.9004329 0.03030303 0.069264069 0.9264069 0.02164502
#> 9  Kayah       Wealthy 0.8494624 0.04301075 0.107526882 0.8387097 0.06451613
#> 11 Kayah        Medium 0.6746411 0.11483254 0.210526316 0.7368421 0.12440191
#> 13 Kayah          Poor 0.6479592 0.12755102 0.224489796 0.7193878 0.14795918
#> 15 Kayah       Poorest 0.7024390 0.19512195 0.102439024 0.7170732 0.20975610
#> 2  Kayin         Rural 0.9075145 0.05202312 0.040462428 0.9104046 0.05780347
#> 4  Kayin         Urban 0.9578454 0.03512881 0.007025761 0.9625293 0.03512881
#> 6  Kayin Hard-to-reach 0.5771429 0.18000000 0.242857143 0.6514286 0.15142857
#> 8  Kayin    Wealthiest 0.9760766 0.01913876 0.004784689 0.9856459 0.01435407
#> 10 Kayin       Wealthy 0.9134199 0.04329004 0.043290043 0.9350649 0.04329004
#> 12 Kayin        Medium 0.7731481 0.11111111 0.115740741 0.8055556 0.09259259
#> 14 Kayin          Poor 0.7149321 0.14932127 0.135746606 0.7601810 0.13122172
#> 16 Kayin       Poorest 0.7264574 0.11210762 0.161434978 0.7488789 0.11659193
#>         rain5   winter1    winter6     winter5
#> 1  0.09090909 0.7492163 0.07523511 0.175548589
#> 3  0.05523256 0.8982558 0.04941860 0.052325581
#> 5  0.14246575 0.6410959 0.16712329 0.191780822
#> 7  0.05194805 0.9047619 0.02597403 0.069264069
#> 9  0.09677419 0.8548387 0.03763441 0.107526882
#> 11 0.13875598 0.6698565 0.11961722 0.210526316
#> 13 0.13265306 0.6581633 0.12755102 0.214285714
#> 15 0.07317073 0.7024390 0.19024390 0.107317073
#> 2  0.03179191 0.9104046 0.05202312 0.037572254
#> 4  0.00234192 0.9601874 0.03512881 0.004683841
#> 6  0.19714286 0.6000000 0.16857143 0.231428571
#> 8  0.00000000 0.9712919 0.02392344 0.004784689
#> 10 0.02164502 0.9177489 0.03896104 0.043290043
#> 12 0.10185185 0.7824074 0.10648148 0.111111111
#> 14 0.10859729 0.7285068 0.14479638 0.126696833
#> 16 0.13452915 0.7488789 0.10313901 0.147982063