Function to create weighted anthropometric indicator results tables

Function to create weighted anthropometric indicator results tables

create_weighted_anthro(df, vars, labs, state, digits = NULL)

Arguments

df
A recoded data.frame of indicators

vars
Variable names for anthropometric indicator/s to create a results table for

labs
Variable labels for anthropometric indicator/s to create a results table for

state
A character value for name of state to report results on. Choice between Kayah or Kayin.

digits
Integer value for number of digits to round off values. Default is NULL for no rounding off.

Value

A data.frame of anthropometric indicator results for each state in total and disaggregated by urban, rural and hard-to-reach areas

Examples

## Recode stunting indicators
canthro  <- recode_anthro(df = create_canthro(df = childAnthro, x = anthroDF))

#> ================================================================================
#> ================================================================================
#> ================================================================================

create_weighted_anthro(df = canthro,
                       vars = c("haz", "global.haz", "moderate.haz", "severe.haz"),
                       labs = c("Mean height-for-age z-score",
                                "Global stunting/stuntedness",
                                "Moderate stunting/stuntedness",
                                "Severe stunting/stuntedness"),
                       state = "Kayah")

#>    state        strata     variable                     indicator    n
#> 1  Kayah         Urban          haz   Mean height-for-age z-score  397
#> 2  Kayah         Rural          haz   Mean height-for-age z-score  411
#> 3  Kayah Hard-to-reach          haz   Mean height-for-age z-score  525
#> 4  Kayah         Total          haz   Mean height-for-age z-score 1333
#> 5  Kayah         Urban   global.haz   Global stunting/stuntedness  390
#> 6  Kayah         Rural   global.haz   Global stunting/stuntedness  409
#> 7  Kayah Hard-to-reach   global.haz   Global stunting/stuntedness  509
#> 8  Kayah         Total   global.haz   Global stunting/stuntedness 1308
#> 9  Kayah         Urban moderate.haz Moderate stunting/stuntedness  390
#> 10 Kayah         Rural moderate.haz Moderate stunting/stuntedness  409
#> 11 Kayah Hard-to-reach moderate.haz Moderate stunting/stuntedness  509
#> 12 Kayah         Total moderate.haz Moderate stunting/stuntedness 1308
#> 13 Kayah         Urban   severe.haz   Severe stunting/stuntedness  390
#> 14 Kayah         Rural   severe.haz   Severe stunting/stuntedness  409
#> 15 Kayah Hard-to-reach   severe.haz   Severe stunting/stuntedness  509
#> 16 Kayah         Total   severe.haz   Severe stunting/stuntedness 1308
#>       estimate          se          lcl         ucl
#> 1  -1.05871537 0.066617917 -1.189284084 -0.92814665
#> 2  -1.39768856 0.075605902 -1.545873409 -1.24950372
#> 3  -1.60982857 0.091990673 -1.790126978 -1.42953017
#> 4  -1.38028507 0.050808549 -1.479867998 -1.28070214
#> 5   0.17435897 0.018907593  0.137300774  0.21141717
#> 6   0.31540342 0.023465309  0.269412262  0.36139458
#> 7   0.47937132 0.030717773  0.419165588  0.53957704
#> 8   0.33715596 0.016908965  0.304015002  0.37029692
#> 9   0.14871795 0.018539449  0.112381297  0.18505460
#> 10  0.25672372 0.022532624  0.212560584  0.30088685
#> 11  0.30058939 0.019273553  0.262813922  0.33836486
#> 12  0.24159021 0.011911570  0.218243965  0.26493646
#> 13  0.02564103 0.008503669  0.008974141  0.04230791
#> 14  0.05867971 0.011519318  0.036102258  0.08125716
#> 15  0.17878193 0.021841956  0.135972479  0.22159137
#> 16  0.09556575 0.010964362  0.074075994  0.11705550