Function to create weighted indicator results tables

Function to create weighted indicator results tables

create_weighted_table(
  df,
  ppiDF = recode_ppi(df = hh),
  vars,
  labs,
  state,
  digits = NULL
)

Arguments

df
A recoded data.frame of indicators

ppiDF
A data.frame produced when recode_ppi is applied to the hh dataset of the Myanmar MCCT Programme Evaluation

vars
Variable names for indicator/s to create a results table for

labs
Variable labels for indicator/s to create a results table for

state
A character value for name of state to report results on. Choice between Kayah or Kayin.

digits
Integer value for number of digits to round off values. Default is NULL for no rounding off.

Value

A data.frame of indicator results for each state in total and disaggregated by urban, rural and hard-to-reach areas and by wealth quintile.

Examples

## Recode MCCT indicators
mcctDF  <- recode_mcct(df = create_mcct(df = hh, x = hhMembers))

create_weighted_table(df = mcctDF,
                      vars = c("mcct1", "mcct2"),
                      labs = c("MCCT coverage",
                               "Mean number of cash transfers received"),
                      state = "Kayah")

#>    state        strata variable                              indicator   n
#> 1  Kayah         Urban    mcct1                          MCCT coverage  90
#> 2  Kayah         Rural    mcct1                          MCCT coverage  93
#> 3  Kayah Hard-to-reach    mcct1                          MCCT coverage 158
#> 4  Kayah         Total    mcct1                          MCCT coverage 341
#> 5  Kayah    Wealthiest    mcct1                          MCCT coverage  63
#> 6  Kayah       Wealthy    mcct1                          MCCT coverage  67
#> 7  Kayah        Medium    mcct1                          MCCT coverage  69
#> 8  Kayah          Poor    mcct1                          MCCT coverage  72
#> 9  Kayah       Poorest    mcct1                          MCCT coverage  63
#> 10 Kayah         Urban    mcct2 Mean number of cash transfers received  62
#> 11 Kayah         Rural    mcct2 Mean number of cash transfers received  69
#> 12 Kayah Hard-to-reach    mcct2 Mean number of cash transfers received  61
#> 13 Kayah         Total    mcct2 Mean number of cash transfers received 192
#> 14 Kayah    Wealthiest    mcct2 Mean number of cash transfers received  45
#> 15 Kayah       Wealthy    mcct2 Mean number of cash transfers received  44
#> 16 Kayah        Medium    mcct2 Mean number of cash transfers received  45
#> 17 Kayah          Poor    mcct2 Mean number of cash transfers received  25
#> 18 Kayah       Poorest    mcct2 Mean number of cash transfers received  29
#>     estimate         se       lcl       ucl
#> 1  0.6888889 0.04806087 0.5946913 0.7830865
#> 2  0.7419355 0.04677270 0.6502627 0.8336083
#> 3  0.3860759 0.06043106 0.2676333 0.5045186
#> 4  0.5630499 0.03491988 0.4946081 0.6314916
#> 5  0.7142857 0.05874691 0.5991439 0.8294275
#> 6  0.6567164 0.06001225 0.5390946 0.7743383
#> 7  0.6521739 0.05668718 0.5410691 0.7632787
#> 8  0.3472222 0.06051633 0.2286124 0.4658320
#> 9  0.4603175 0.08471293 0.2942832 0.6263517
#> 10 2.1129032 0.12589451 1.8661545 2.3596519
#> 11 2.2028986 0.14185475 1.9248683 2.4809288
#> 12 2.0655738 0.18355457 1.7058134 2.4253341
#> 13 2.1302083 0.08725885 1.9591841 2.3012325
#> 14 2.1555556 0.10673585 1.9463571 2.3647540
#> 15 2.1363636 0.16818789 1.8067214 2.4660058
#> 16 2.2888889 0.14521243 2.0042778 2.5735000
#> 17 1.9600000 0.20888747 1.5505881 2.3694119
#> 18 1.9655172 0.20215488 1.5693010 2.3617335