Function to get number of samples used to calculate indicator stratified by geographic area

Function to get number of samples used to calculate indicator stratified by geographic area

get_n_area(df, vars)

Arguments

df
A data.frame of recoded indicator variables

vars
A character vector of variable names to get number of samples for

Value

A table of numeric values of number of samples use to calculate specific indicators by area strata

Examples

get_n_area(df = recode_anthro(df = create_canthro(df = childAnthro,
                                                  x = anthroDF)),
           vars = c("global.haz", "moderate.haz", "severe.haz"))

#> ================================================================================
#> ================================================================================
#> ================================================================================

#> $Kayah
#>               global.haz moderate.haz severe.haz
#> Urban                390          390        390
#> Rural                409          409        409
#> Hard-to-reach        509          509        509
#> Total               1308         1308       1308
#> 
#> $Kayin
#>               global.haz moderate.haz severe.haz
#> Urban                524          524        524
#> Rural                499          499        499
#> Hard-to-reach        515          515        515
#> Total               1538         1538       1538
#>