Function to recode early initiation of breastfeeding indicator for the Myanmar MCCT Programme Evaluation Survey

Function to recode early initiation of breastfeeding indicator for the Myanmar MCCT Programme Evaluation Survey

recode_eibf(df)

Arguments

df
A data.frame collected for the Myanmar MCCT Programme Evaluation Study containing IYCF/breastfeeding data

Value

A vector of recoded early initiation of breastfeeding status

Examples

iycfDF <- create_iycf(df = iycf, x = hh, y = hhMembers)
recode_eibf(df = iycfDF)

#>    [1] NA  0  1  1  1 NA  1 NA NA  1  1  1 NA NA  1  1  1  1 NA  1 NA  1  1 NA
#>   [25]  0  1  0 NA  0  1 NA  1  0  1  1  1  1 NA NA  1 NA  1 NA  0 NA  1  1 NA
#>   [49]  1  1  1 NA  0 NA  1  1  1 NA  1  1  1 NA  1  1  0  1  1  1  1  1  1  1
#>   [73]  1  1  1 NA  0  1  1  1  1  1  1  1  1  0  1  1  1  1  1  1  1  1  1  1
#>   [97]  1  1  1  1  1  1  1  1  1  1  0  1  0 NA  1  0  1  1  1  1 NA  1  1  1
#>  [121] NA  1  1  0  0  1 NA  1  0  1  1 NA  1  1 NA  1 NA  1  1  1  1  1  1  1
#>  [145]  1  1 NA NA NA NA  1  1  1  1 NA NA NA NA  1  1  1 NA  1  0  0  1  1  1
#>  [169]  1  1  1 NA  1  1  1  1  1  1 NA  1 NA  1  1  1 NA  1 NA  1  1  1 NA  0
#>  [193]  1  1  1  0  0  1 NA  1  1  1  1  0  1  1  1 NA  1  1  1  1  0  1  0  1
#>  [217] NA NA  1  1  1 NA  0  1  1  1 NA  0  1  1  1  1 NA NA  1  1 NA  1  1 NA
#>  [241]  1  1  1 NA  1  1  1  1  1  1  1 NA NA NA NA  1  1  1 NA NA NA  1 NA  1
#>  [265]  1  0  1  1 NA  0 NA  1  1  1  1  1 NA  0 NA  1 NA NA NA NA  0  1  1 NA
#>  [289] NA  1  1  0  1  1  1  1 NA  1  1 NA  0 NA  1  1  1  1  1 NA  1  1  1  0
#>  [313]  1  1  1  0  1  1  1  1 NA  1  1  1  0  0 NA  1 NA  1  1 NA  1  0  1  1
#>  [337]  1  1  1 NA NA  1 NA  1 NA NA  1 NA  1  1  1 NA NA  1 NA NA  1  1  1  1
#>  [361]  0  1  1  1  1 NA NA NA  1  1 NA  0  1  1  1  1  1  1  1  1  1 NA NA  1
#>  [385]  1  1  1  1  1 NA NA  1  1  1  1 NA  1  1  1  1  0 NA  1 NA NA  0  1  1
#>  [409]  1  1  1  1  1  1  1 NA NA  1  1  1  1  1  1 NA NA  1  0  1  1  1  1  1
#>  [433]  1  1 NA  1 NA NA  1  1  1 NA NA NA  1  1  1  1  1  1  1  1  1  1  1 NA
#>  [457]  1 NA  1  1 NA NA  1  1 NA  1  1  1 NA NA NA  1  1  1 NA NA NA  1  1  1
#>  [481] NA NA  1  1  1  1  1 NA  1  1  1 NA  1  1  1  0  1 NA NA NA  1  1 NA  1
#>  [505] NA  1  1  1  1  1 NA  1 NA NA  1  1  1  1  1 NA  1  1  1  1  1  0 NA  1
#>  [529] NA  0  1  1  1  1  1  0  1  1  1  1  1 NA  1  1  1  1  1  1  1  1  1  1
#>  [553] NA NA  1  1  1  1 NA NA  1  0  0  0  1  1  1 NA  0  1  1  1  1  1  0  0
#>  [577]  0  1  1  0  1 NA NA  1  1 NA  0 NA  1 NA NA  0  0  1  1 NA  1  1  1  1
#>  [601]  1  1  1  1  1  1  1  1  1 NA  1  1  1  1  1  0  1 NA  1  1 NA  1  0  1
#>  [625]  1  1  1  1  1 NA NA  1 NA  1 NA  1 NA  1  1 NA  1  0  1  1  1  1  1  1
#>  [649]  1 NA  0  1  1  1  1  1  1  1 NA  1 NA  1  1  1  0  1  0 NA NA  1 NA NA
#>  [673]  1  1  1  1  1 NA  1  1  0  0  1  1 NA  1 NA  1  1  1 NA  1  0  1  1  1
#>  [697]  1 NA NA  0  1  1  1  1  1  1  1  1  1  0  1  1  1  1  1 NA  1  1  1 NA
#>  [721]  1  1  1  1 NA  1  1 NA NA  1  0  1  1 NA  1  0  1  0  1  1  1  1 NA  1
#>  [745] NA  1  1  1  1  0 NA NA  0 NA  0  1  0 NA  0  1  1 NA NA  1  1  1 NA  1
#>  [769]  0  1  1  0  1 NA  1  1  1 NA  1  1  1  1  1 NA  1 NA NA  0  1  1  1  0
#>  [793]  1  1  0 NA  1 NA  0  1 NA NA  1 NA NA NA  1  1 NA  1  0  1  1  1  1  1
#>  [817] NA  1  0  1  1  1  1  1  1  1  1  1  1 NA NA  1 NA NA  0  1  0 NA  1  0
#>  [841]  1 NA  0 NA  1  1  1 NA NA NA  1  1  1  1 NA  1  0  1 NA NA  1  1  1  1
#>  [865] NA  1 NA  0  1 NA  1  1  0 NA NA  0  1  1  1  1  1  1  1  0  1 NA  1  1
#>  [889]  1  1  1  1  1  1  0  1  1 NA  1 NA  1  1  1 NA  1  0 NA  1  1  1  1 NA
#>  [913] NA NA  1  1  1  1  1  0  1  1  0  0  0 NA NA NA  1  1  1 NA NA NA  1  1
#>  [937] NA  1 NA  1  1  1  1 NA  1  1  1 NA  1  1 NA  1 NA  1  1  0  0  1  1  1
#>  [961] NA NA  1  1  0  1  1  1  1 NA  1  1 NA  1  1 NA  1  1  1  1  1 NA  1 NA
#>  [985] NA NA  1  1  1  1  1 NA  0  1  1  1  1 NA  1  1  1  1 NA  0  1  1  0  0
#> [1009]  1  1  1  1  1  1 NA  1 NA  0 NA  1  1  1  0  1  0 NA  1  1  1  1  0 NA
#> [1033]  1  1  1  1  1  1  1  1  1  1 NA NA  1  1  0  1  1  1  1  1  1  0  0  1
#> [1057]  0  0 NA NA  0 NA  1 NA  1  0  1  1  1  0  0  0  1  1  1  0  1  1  1  1
#> [1081]  1  1  1  1  1  0  1  1 NA NA  0  1  1  1  1  1  0  0  1  1 NA  1  1  1
#> [1105] NA  1 NA  1 NA NA  1 NA NA  1  1 NA NA NA  1  0  1  1  1  1  1 NA  1  1
#> [1129]  1  1 NA  1  0  0  1  1  1  1  1  1  1  0  1  1 NA  0  1  0 NA  1 NA  1
#> [1153]  1  1  1  1 NA NA  1 NA  1 NA  1  0 NA NA  1  1  1 NA  1  1  1  1  1  1
#> [1177]  0 NA NA  1 NA  1  1  1 NA NA NA  1 NA  1  1  1  1  1  1  1  1 NA  0  0
#> [1201]  1  0  1  0  1 NA NA  1 NA  1  1  1  1  1 NA  1  1  1  1 NA  1  1  1  1
#> [1225]  1  1  0  1  1 NA  0  1  1 NA NA  1  1  1  1 NA  1  1 NA  1  0  1  1  1
#> [1249]  1  1 NA  1  1  1  1  1  1  1  1  1 NA  1  1  1  1  1  1  1  1  1  1  0
#> [1273]  1  1  0  0  1  0  1  0  1  1 NA  0  1  1  1  0  1 NA  1  1  0  1  1 NA
#> [1297]  1  1  1  1 NA NA NA  1 NA NA NA  1  1  1  1  1  1  1  1  0  1 NA  1  1
#> [1321]  1 NA  1  1 NA  1  0  1  1  1 NA  1  1  1  0 NA  1 NA NA  1  0 NA  1  1
#> [1345]  1  1  1  1 NA  0 NA  0  1  1 NA  0 NA  1  1  1  1  1 NA NA  1  1  1  1
#> [1369] NA  0 NA  1  1  1 NA  1  1 NA  1  1  1  0  1  1  1 NA  1  0  1  0  0  1
#> [1393]  1 NA  1  0  1  1  1  1  1  1  1  0  0  1 NA  1  1  1  1  1  1 NA  1  1
#> [1417] NA  1  1  1 NA  1  1 NA  0 NA NA  1  1 NA NA NA NA NA  1  1  1 NA  1 NA
#> [1441] NA  1  1  1  1  1  1  1  1  1 NA  1  1  1  1  1 NA  1  1  1 NA  1  1  0
#> [1465]  1  1  1 NA NA  1  0 NA  1  1 NA  1  1 NA  1  1  1  1  1  1 NA  1 NA  1
#> [1489]  1 NA  1  1  1 NA  1 NA  0  1  1  1  1  1  1  0  1  1  1  1  0  1 NA  0
#> [1513]  1  0  1  1  1 NA  1 NA  1 NA  1 NA NA NA  1  1  1 NA  1  1  1 NA  1  1
#> [1537]  1  1 NA  1  0  1  1  1 NA  1 NA NA NA  1  1  1  1  1  1  0  1  1 NA  1
#> [1561]  1  1  0  1  1  1  1  1  1  1  1 NA NA  1  0  0  1  1  1  1  0 NA  1  1
#> [1585]  1  1  1  1  1 NA  1  1  1  1  1 NA NA  0 NA  1  1 NA  1 NA  1  0  1 NA
#> [1609]  1  1  1  1  1  1  1  1  0  1  1  0  1  1  1  1  1  1  1  1  1  1  1  1
#> [1633]  0  1  1 NA  1  1  1  1  1  1  1 NA  1 NA  1  1  1