Function to recode early initiation of breastfeeding indicator for the Myanmar MCCT Programme Evaluation Survey

Function to recode early initiation of breastfeeding indicator for the Myanmar MCCT Programme Evaluation Survey

recode_eibf(df)

Arguments

df
A data.frame collected for the Myanmar MCCT Programme Evaluation Study containing IYCF/breastfeeding data

Value

A vector of recoded early initiation of breastfeeding status

Examples

iycfDF <- create_iycf(df = iycf, x = hh, y = hhMembers)
recode_eibf(df = iycfDF)

#>    [1] NA  0  1  1  1 NA  1 NA NA  1  1  1 NA NA  1  1  1  1 NA  1  1 NA  1  1
#>   [25] NA  0  1  0 NA  0  1 NA  1  0  1  1  1  1 NA NA  1 NA  1 NA  0 NA  1  1
#>   [49] NA  1  1  1 NA  0 NA  1  1  1 NA  1  1  1 NA  1  1  0  1  1  1  1  1  1
#>   [73]  1  1  1  1 NA  0  1  1  1  1  1  1  1  1  0  1  1  1  1  1  1  1  1  1
#>   [97]  1  1  1  1  1  1  1  1  1  1  1  0  1  0 NA  1  0  1  1  1  1 NA  1  1
#>  [121]  1 NA  1  1  0  0  1 NA  1  0  1  1 NA  1  1 NA  1 NA  1  1  1  1  1  1
#>  [145]  1  1  1 NA NA NA NA  1  1  1  1 NA NA NA NA  1  1  1 NA  1  0  0  1  1
#>  [169]  1  1  1 NA  1  1  1  1  1  1  1 NA  1 NA  1  1  1 NA  1 NA  1  1  1 NA
#>  [193]  0  1  1  1  0  0  1 NA  1  1  1  1  0  1  1  1 NA  1  1  1  1  0  1  0
#>  [217]  1 NA NA  1  1  1 NA  0  1  1  1 NA  0  1  1  1  1 NA NA  1  1 NA  1  1
#>  [241] NA  1  1  1 NA  1  1  1  1  1  1  1 NA NA NA NA  1  1  1 NA NA NA  1 NA
#>  [265]  1  1  0  1  1 NA  0 NA  1  1  1  1  1 NA  0 NA  1 NA NA NA NA  0  1  1
#>  [289] NA NA  1  1  0  1  1  1  1 NA  1  1 NA  0 NA  1  1  1  1  1 NA  1  1  1
#>  [313]  0  1  1  1  0  1  1  1  1 NA  1  1  1  0  0 NA  1 NA  1  1 NA  1  0  1
#>  [337]  1  1  1  1 NA NA  1 NA  1 NA NA  1 NA  1  1  1 NA NA  1 NA NA  1  1  1
#>  [361]  1  0  1  1  1  1 NA NA NA  1  1 NA  0  1  1  1  1  1  1  1  1  1 NA NA
#>  [385]  1  1  1  1  1  1 NA NA  1  1  1  1 NA  1  1  1  1  0 NA  1 NA NA  0  1
#>  [409]  1  1  1  1  1  1  1  1 NA NA  1  1  1  1  1  1 NA NA  1  0  1  1  1  1
#>  [433]  1  1  1 NA  1 NA NA  1  1  1 NA NA NA  1  1  1  1  1  1  1  1  1  1  1
#>  [457] NA  1 NA  1  1 NA NA  1  1 NA  1  1  1 NA NA NA  1  1  1 NA NA NA  1  1
#>  [481]  1 NA NA  1  1  1  1  1 NA  1  1  1 NA  1  1  1  0  1 NA NA NA  1  1 NA
#>  [505]  1 NA  1  1  1  1  1 NA  1 NA NA  1  1  1  1  1 NA  1  1  1  1  1  0 NA
#>  [529]  1 NA  0  1  1  1  1  1  0  1  1  1  1  1 NA  1  1  1  1  1  1  1  1  1
#>  [553]  1 NA NA  1  1  1  1 NA NA  1  0  0  0  1  1  1 NA  0  1  1  1  1  1  0
#>  [577]  0  0  1  1  0  1 NA NA  1  1 NA  0 NA  1 NA NA  0  0  1  1 NA  1  1  1
#>  [601]  1  1  1  1  1  1  1  1  1  1 NA  1  1  1  1  1  0  1 NA  1  1 NA  1  0
#>  [625]  1  1  1  1  1  1 NA NA  1 NA  1 NA  1 NA  1  1 NA  1  0  1  1  1  1  1
#>  [649]  1  1 NA  0  1  1  1  1  1  1  1 NA  1 NA  1  1  1  0  1  0 NA NA  1 NA
#>  [673] NA  1  1  1  1  1 NA  1  1  0  0  1 NA  1  1 NA  1  1  1 NA  1  0  1  1
#>  [697]  1  1 NA NA  0  1  1  1  1  1  1  1  1  1  0  1  1  1  1  1 NA  1  1  1
#>  [721] NA  1  1  1  1 NA  1  1 NA NA  1  0  1  1 NA  1  0  1  0  1  1  1 NA  1
#>  [745]  1 NA  1  1  1  1  0 NA NA  0 NA  0  0  1 NA  0  1  1 NA  1 NA  1  1 NA
#>  [769]  1  0  1  1  0  1 NA  1  1  1 NA  1  1  1  1  1 NA  1 NA NA  0  1  1  1
#>  [793]  0  1  1  0 NA  1 NA  0  1 NA NA  1 NA NA NA  1  1 NA  1  0  1  1  1  1
#>  [817]  1 NA  1  0  1  1  1  1  1  1  1  1  1  1 NA NA  1 NA NA  0  1  0 NA  1
#>  [841]  0  1 NA  0 NA  1  1  1 NA NA NA  1  1  1  1 NA  1  0  1 NA NA  1  1  1
#>  [865]  1 NA  1 NA  0  1 NA  1  1  0 NA NA  0  1  1  1  1  1  1  1  0  1 NA  1
#>  [889]  1  1  1  1  1  1  1  0  1  1 NA  1 NA  1  1  1 NA  1  0 NA  1  1  1  1
#>  [913] NA NA NA  1  1  1  1  1  0  1  1  0  0  0 NA NA NA  1  1  1 NA NA NA  1
#>  [937]  1 NA  1 NA  1  1  1  1 NA  1  1  1 NA  1  1 NA  1 NA  1  1  0  0  1  1
#>  [961]  1 NA NA  1  1  0  1  1  1  1 NA  1  1 NA  1  1 NA  1  1  1  1  1 NA  1
#>  [985] NA NA NA  1  1  1  1  1 NA  0  1  1  1  1 NA  1  1  1  1 NA  0  1  1  0
#> [1009]  0  1  1  1  1  1  1 NA  1 NA  0 NA  1  1  1  0  1  0 NA  1  1  1  1  0
#> [1033] NA  1  1  1  1  1  1  1  1  1  1 NA NA  1  1  0  1  1  1  1  1  1  0  0
#> [1057]  1  0  0 NA NA  0 NA  1 NA  1  0  1  1  1  0  0  0  1  1  1  0  1  1  1
#> [1081]  1  1  1  1  1  1  0  1  1 NA NA  0  1  1  1  1  1  0  0  1  1 NA  1  1
#> [1105]  1 NA NA  1  1 NA NA  1 NA  1 NA  1 NA NA NA  1  0  1  1  1  1  1 NA  1
#> [1129]  1  1  1 NA  1  0  0  1  1  1  1  1  1  1  0  1  1 NA  0  1  0 NA  1 NA
#> [1153]  1  1  1  1  1 NA NA  1 NA  1 NA  1  0 NA NA  1  1  1 NA  1  1  1  1  1
#> [1177]  1  0 NA NA  1 NA  1  1  1 NA NA NA  1 NA  1  1  1  1  1  1  1  1 NA  0
#> [1201]  0  1  0  1  0  1 NA NA  1 NA  1  1  1  1  1 NA  1  1  1  1 NA  1  1  1
#> [1225]  1  1  1  0  1  1 NA  0  1  1 NA NA  1  1  1  1 NA  1  1 NA  1  0  1  1
#> [1249]  1  1  1 NA  1  1  1  1  1  1  1  1  1 NA  1  1  1  1  1  1  1  1  1  1
#> [1273]  0  1  1  0  0  1  0  1  0  1  1 NA  0  1  1  1  0  1 NA  1  1  0  1  1
#> [1297] NA  1  1  1  1 NA NA NA  1 NA NA NA  1  1  1  1  1  1  1  1  0  1 NA  1
#> [1321]  1  1 NA  1  1 NA  1  0  1  1  1 NA  1  1  1  0 NA  1 NA NA  1  0 NA  1
#> [1345]  1  1  1  1  1 NA  0 NA  0  1  1 NA  0 NA  1  1  1  1  1 NA NA  1  1  1
#> [1369]  1 NA  0 NA  1  1  1 NA  1  1 NA  1  1  1  0  1  1  1 NA  1  0  1  0  0
#> [1393]  1  1 NA  1  0  1  1  1  1  1  1  1  0  0  1 NA  1  1  1  1  1  1 NA  1
#> [1417]  1 NA  1  1  1 NA  1  1 NA  0 NA NA  1  1 NA NA NA NA NA  1  1  1 NA  1
#> [1441] NA NA  1  1  1  1  1  1  1  1  1 NA  1  1  1  1  1 NA  1  1  1 NA  1  1
#> [1465]  0  1  1  1 NA NA  1  0 NA  1  1 NA  1  1 NA  1  1  1  1  1  1 NA  1 NA
#> [1489]  1  1 NA  1  1  1 NA  1 NA  0  1  1  1  1  1  1  0  1  1  1  1  0  1 NA
#> [1513]  0  1  0  1  1  1 NA  1 NA  1 NA  1 NA NA NA  1  1  1 NA  1  1  1 NA  1
#> [1537]  1  1  1 NA  1  0  1  1  1 NA  1 NA NA NA  1  1  1  1  1  1  0  1  1 NA
#> [1561]  1  1  1  0  1  1  1  1  1  1  1  1 NA NA  1  0  0  1  1  1  1  0 NA  1
#> [1585]  1  1  1  1  1  1 NA  1  1  1  1  1 NA NA  0 NA  1  1 NA  1 NA  1  0  1
#> [1609] NA  1  1  1  1  1  1  1  1  0  1  1  0  1  1  1  1  1  1  1  1  1  1  1
#> [1633]  1  0  1  1 NA  1  1  1  1  1  1  1 NA  1 NA  1  1  1